PROD-2211 moves some property-specific endpoints from OSS -> plus (#5069)
